requirements.txt
requests>=2.28.1
dnspython>=2.2.1
colorama>=0.4.6
argparse>=1.4.0